introduce-raw-tokenizer-wrapper(helpers-move): Hand tokenize-pipeline helpers over to RawTokenizerWrapper
Pure cut/paste move per MECH_COMMIT_SPLIT. Move the tokenize-pipeline
helpers (``_detect_input_format`` / ``_prepare_tokenizer_input`` /
``_extract_tokenizer_results`` / ``_tokenize_texts``) and the
``InputFormat`` enum from TokenizerManager into RawTokenizerWrapper (drop
@staticmethod, restore plain ``self``, collapse intra-cluster
``TokenizerManager._x(self, ...)`` calls back to ``self._x(...)`` — they
are sibling instance methods on the new owner).

External caller prefix replacement in TM:
``TokenizerManager._helper(self.raw_tokenizer_wrapper, ...)`` →
``self.raw_tokenizer_wrapper._helper(...)``. Update the manual test
imports and call sites to reach the helpers and InputFormat through
raw_tokenizer_wrapper.

Refactor chain ID: rtw-move-tokenize-helpers
PR-Title: Extract RawTokenizerWrapper from TokenizerManager
