+++ python/sglang/srt/managers/tokenizer_manager.py
@@ -140,14 +140,6 @@ _REQUEST_STATE_WAIT_TIMEOUT = envs.SGLANG_REQUEST_STATE_WAIT_TIMEOUT.get()
 logger = logging.getLogger(__name__)
 
 
-class InputFormat(Enum):
-    """Input format types for tokenization handling."""
-
-    SINGLE_STRING = 1  # Regular single text like "Hello world"
-    BATCH_STRINGS = 2  # Regular batch like ["Hello", "World"]
-    CROSS_ENCODER_PAIRS = 3  # Cross-encoder pairs like [["query", "document"]]
-
-
 class TokenizerManager(TokenizerControlMixin):
     """TokenizerManager is a process that tokenizes the text."""
 
@@ -494,165 +486,6 @@ class TokenizerManager(TokenizerControlMixin):
                 async for response in self._handle_batch_request(obj, request):
                     yield response
 
-    @staticmethod
-    def _detect_input_format(
-        self: "RawTokenizerWrapper",
-        texts: Union[str, List[str]],
-        is_cross_encoder: bool,
-    ) -> InputFormat:
-        """Detect the format of input texts for proper tokenization handling.
-
-        Returns:
-            - InputFormat.SINGLE_STRING: Regular single text like "Hello world"
-            - InputFormat.BATCH_STRINGS: Regular batch like ["Hello", "World"]
-            - InputFormat.CROSS_ENCODER_PAIRS: Cross-encoder pairs like [["query", "document"]]
-        """
-        if isinstance(texts, str):
-            return InputFormat.SINGLE_STRING
-
-        if (
-            is_cross_encoder
-            and len(texts) > 0
-            and isinstance(texts[0], list)
-            and len(texts[0]) == 2
-        ):
-            return InputFormat.CROSS_ENCODER_PAIRS
-
-        return InputFormat.BATCH_STRINGS
-
-    @staticmethod
-    def _prepare_tokenizer_input(
-        self: "RawTokenizerWrapper",
-        texts: Union[str, List[str]],
-        input_format: InputFormat,
-    ) -> Union[List[str], List[List[str]]]:
-        """Prepare input for the tokenizer based on detected format."""
-        if input_format == InputFormat.SINGLE_STRING:
-            return [texts]  # Wrap single string for batch processing
-        elif input_format == InputFormat.CROSS_ENCODER_PAIRS:
-            return texts  # Already in correct format: [["query", "doc"]]
-        else:  # BATCH_STRINGS
-            return texts  # Already in correct format: ["text1", "text2"]
-
-    @staticmethod
-    def _extract_tokenizer_results(
-        self: "RawTokenizerWrapper",
-        input_ids: List[List[int]],
-        token_type_ids: Optional[List[List[int]]],
-        input_format: InputFormat,
-        original_batch_size: int,
-    ) -> Union[
-        Tuple[List[int], Optional[List[int]]],
-        Tuple[List[List[int]], Optional[List[List[int]]]],
-    ]:
-        """Extract results from tokenizer output based on input format."""
-
-        # For single inputs (string or single cross-encoder pair), extract first element
-        if (
-            input_format in [InputFormat.SINGLE_STRING, InputFormat.CROSS_ENCODER_PAIRS]
-            and original_batch_size == 1
-        ):
-            single_input_ids = input_ids[0] if input_ids else []
-            single_token_type_ids = token_type_ids[0] if token_type_ids else None
-            return single_input_ids, single_token_type_ids
-
-        # For true batches, return as-is
-        return input_ids, token_type_ids
-
-    @staticmethod
-    async def _tokenize_texts(
-        self: "RawTokenizerWrapper",
-        texts: Union[str, List[str]],
-        is_cross_encoder: bool = False,
-    ) -> Union[
-        Tuple[List[int], Optional[List[int]]],
-        Tuple[List[List[int]], Optional[List[List[int]]]],
-    ]:
-        """
-        Tokenize text(s) using the appropriate tokenizer strategy.
-
-        This method handles multiple input formats and chooses between async dynamic
-        batch tokenizer (for single texts only) and regular tokenizer.
-
-        Args:
-            texts: Text input in various formats:
-
-                   Regular cases:
-                   - Single string: "How are you?"
-                   - Batch of strings: ["Hello", "World", "How are you?"]
-
-                   Cross-encoder cases (sentence pairs for similarity/ranking):
-                   - Single pair: [["query text", "document text"]]
-                   - Multiple pairs: [["q1", "d1"], ["q2", "d2"], ["q3", "d3"]]
-
-            is_cross_encoder: Whether to return token_type_ids for cross-encoder models.
-                             Enables proper handling of sentence pairs with segment IDs.
-
-        Returns:
-            Single input cases:
-                Tuple[List[int], Optional[List[int]]]: (input_ids, token_type_ids)
-                Example: ([101, 2129, 102], [0, 0, 0]) for single text
-                Example: ([101, 2129, 102, 4068, 102], [0, 0, 0, 1, 1]) for cross-encoder pair
-
-            Batch input cases:
-                Tuple[List[List[int]], Optional[List[List[int]]]]: (batch_input_ids, batch_token_type_ids)
-                Example: ([[101, 2129, 102], [101, 4068, 102]], None) for regular batch
-
-            Note: token_type_ids is None unless is_cross_encoder=True.
-        """
-        if not texts or self.tokenizer is None:
-            raise ValueError("texts cannot be empty and tokenizer must be initialized")
-
-        # Step 1: Detect input format and prepare for tokenization
-        input_format = TokenizerManager._detect_input_format(
-            self, texts, is_cross_encoder
-        )
-        tokenizer_input = TokenizerManager._prepare_tokenizer_input(
-            self, texts, input_format
-        )
-        original_batch_size = len(texts) if not isinstance(texts, str) else 1
-
-        # Step 2: Set up tokenizer arguments
-        tokenizer_kwargs = (
-            {"return_token_type_ids": is_cross_encoder} if is_cross_encoder else {}
-        )
-
-        # Step 3: Choose tokenization strategy
-        use_async_tokenizer = (
-            self.async_dynamic_batch_tokenizer is not None
-            and input_format == InputFormat.SINGLE_STRING
-        )
-
-        if use_async_tokenizer:
-            logger.debug("Using async dynamic batch tokenizer for single text")
-            result = await self.async_dynamic_batch_tokenizer.encode(
-                tokenizer_input[0], **tokenizer_kwargs
-            )
-            # Convert to batch format for consistency
-            input_ids = [result["input_ids"]]
-            token_type_ids = (
-                [result["token_type_ids"]]
-                if is_cross_encoder and result.get("token_type_ids")
-                else None
-            )
-        else:
-            logger.debug(f"Using regular tokenizer for {len(tokenizer_input)} inputs")
-
-            if not is_cross_encoder and (not getattr(self.tokenizer, "is_fast", False)):
-                input_ids = [self.tokenizer.encode(t) for t in tokenizer_input]
-                token_type_ids = None
-            else:
-                encoded = self.tokenizer(tokenizer_input, **tokenizer_kwargs)
-                input_ids = encoded["input_ids"]
-                token_type_ids = (
-                    encoded.get("token_type_ids") if is_cross_encoder else None
-                )
-
-        # Step 4: Extract results based on input format
-        return TokenizerManager._extract_tokenizer_results(
-            self, input_ids, token_type_ids, input_format, original_batch_size
-        )
-
     async def _tokenize_one_request(
         self,
         obj: Union[GenerateReqInput, EmbeddingReqInput],
@@ -690,8 +523,10 @@ class TokenizerManager(TokenizerControlMixin):
                 # Use empty placeholder - multimodal processor will override
                 input_ids = []
             else:
-                input_ids, token_type_ids = await TokenizerManager._tokenize_texts(
-                    self.raw_tokenizer_wrapper, input_text, is_cross_encoder_request
+                input_ids, token_type_ids = (
+                    await self.raw_tokenizer_wrapper._tokenize_texts(
+                        input_text, is_cross_encoder_request
+                    )
                 )
 
         contains_mm_input = obj.contains_mm_input()
@@ -1103,8 +938,10 @@ class TokenizerManager(TokenizerControlMixin):
         )
 
         # Batch tokenize all texts using unified method
-        input_ids_list, token_type_ids_list = await TokenizerManager._tokenize_texts(
-            self.raw_tokenizer_wrapper, texts, is_cross_encoder_request
+        input_ids_list, token_type_ids_list = (
+            await self.raw_tokenizer_wrapper._tokenize_texts(
+                texts, is_cross_encoder_request
+            )
         )
 
         # Process all requests
