+++ test/manual/test_tokenizer_manager.py
@@ -18,10 +18,9 @@ import unittest
 from unittest.mock import Mock, patch
 
 from sglang.srt.managers.io_struct import GenerateReqInput
-from sglang.srt.managers.tokenizer_manager import (
+from sglang.srt.managers.tokenizer_manager import ReqState, TokenizerManager
+from sglang.srt.managers.tokenizer_manager_components.raw_tokenizer_wrapper import (
     InputFormat,
-    ReqState,
-    TokenizerManager,
 )
 from sglang.srt.observability.req_time_stats import APIServerReqTimeStats
 from sglang.srt.server_args import PortArgs, ServerArgs
@@ -50,78 +49,78 @@ class TestInputFormatDetection(unittest.TestCase):
     def test_detect_single_string(self):
         """Test detection of single string input."""
         text = "Hello world"
-        result = TokenizerManager._detect_input_format(
-            self.tokenizer_manager.raw_tokenizer_wrapper, text, is_cross_encoder=False
+        result = self.tokenizer_manager.raw_tokenizer_wrapper._detect_input_format(
+            text, is_cross_encoder=False
         )
         self.assertEqual(result, InputFormat.SINGLE_STRING)
 
     def test_detect_single_string_cross_encoder_disabled(self):
         """Test single string with cross_encoder disabled still returns single_string."""
         text = "Hello world"
-        result = TokenizerManager._detect_input_format(
-            self.tokenizer_manager.raw_tokenizer_wrapper, text, is_cross_encoder=True
+        result = self.tokenizer_manager.raw_tokenizer_wrapper._detect_input_format(
+            text, is_cross_encoder=True
         )
         self.assertEqual(result, InputFormat.SINGLE_STRING)
 
     def test_detect_batch_strings(self):
         """Test detection of batch string inputs."""
         texts = ["Hello", "World", "How are you?"]
-        result = TokenizerManager._detect_input_format(
-            self.tokenizer_manager.raw_tokenizer_wrapper, texts, is_cross_encoder=False
+        result = self.tokenizer_manager.raw_tokenizer_wrapper._detect_input_format(
+            texts, is_cross_encoder=False
         )
         self.assertEqual(result, InputFormat.BATCH_STRINGS)
 
     def test_detect_batch_strings_cross_encoder_disabled(self):
         """Test batch strings with cross_encoder disabled."""
         texts = ["Hello", "World"]
-        result = TokenizerManager._detect_input_format(
-            self.tokenizer_manager.raw_tokenizer_wrapper, texts, is_cross_encoder=True
+        result = self.tokenizer_manager.raw_tokenizer_wrapper._detect_input_format(
+            texts, is_cross_encoder=True
         )
         self.assertEqual(result, InputFormat.BATCH_STRINGS)
 
     def test_detect_cross_encoder_single_pair(self):
         """Test detection of cross-encoder single pair."""
         texts = [["query text", "document text"]]
-        result = TokenizerManager._detect_input_format(
-            self.tokenizer_manager.raw_tokenizer_wrapper, texts, is_cross_encoder=True
+        result = self.tokenizer_manager.raw_tokenizer_wrapper._detect_input_format(
+            texts, is_cross_encoder=True
         )
         self.assertEqual(result, InputFormat.CROSS_ENCODER_PAIRS)
 
     def test_detect_cross_encoder_multiple_pairs(self):
         """Test detection of cross-encoder multiple pairs."""
         texts = [["q1", "d1"], ["q2", "d2"], ["q3", "d3"]]
-        result = TokenizerManager._detect_input_format(
-            self.tokenizer_manager.raw_tokenizer_wrapper, texts, is_cross_encoder=True
+        result = self.tokenizer_manager.raw_tokenizer_wrapper._detect_input_format(
+            texts, is_cross_encoder=True
         )
         self.assertEqual(result, InputFormat.CROSS_ENCODER_PAIRS)
 
     def test_detect_cross_encoder_disabled_with_pairs(self):
         """Test pairs with cross_encoder disabled should return batch_strings."""
         texts = [["query", "document"]]
-        result = TokenizerManager._detect_input_format(
-            self.tokenizer_manager.raw_tokenizer_wrapper, texts, is_cross_encoder=False
+        result = self.tokenizer_manager.raw_tokenizer_wrapper._detect_input_format(
+            texts, is_cross_encoder=False
         )
         self.assertEqual(result, InputFormat.BATCH_STRINGS)
 
     def test_detect_empty_list(self):
         """Test detection with empty list."""
         texts = []
-        result = TokenizerManager._detect_input_format(
-            self.tokenizer_manager.raw_tokenizer_wrapper, texts, is_cross_encoder=True
+        result = self.tokenizer_manager.raw_tokenizer_wrapper._detect_input_format(
+            texts, is_cross_encoder=True
         )
         self.assertEqual(result, InputFormat.BATCH_STRINGS)
 
     def test_detect_malformed_cross_encoder_pairs(self):
         """Test malformed cross-encoder pairs (not length 2)."""
         texts = [["query only"]]  # Single element, not a pair
-        result = TokenizerManager._detect_input_format(
-            self.tokenizer_manager.raw_tokenizer_wrapper, texts, is_cross_encoder=True
+        result = self.tokenizer_manager.raw_tokenizer_wrapper._detect_input_format(
+            texts, is_cross_encoder=True
         )
         self.assertEqual(result, InputFormat.BATCH_STRINGS)
 
         texts = [["query", "doc", "extra"]]  # Three elements, not a pair
-        result = TokenizerManager._detect_input_format(
-            self.tokenizer_manager.raw_tokenizer_wrapper, texts, is_cross_encoder=True
+        result = self.tokenizer_manager.raw_tokenizer_wrapper._detect_input_format(
+            texts, is_cross_encoder=True
         )
         self.assertEqual(result, InputFormat.BATCH_STRINGS)
 
@@ -148,8 +147,7 @@ class TestTokenizerInputPreparation(unittest.TestCase):
     def test_prepare_single_string_input(self):
         """Test preparation of single string input."""
         text = "Hello world"
-        result = TokenizerManager._prepare_tokenizer_input(
-            self.tokenizer_manager.raw_tokenizer_wrapper,
+        result = self.tokenizer_manager.raw_tokenizer_wrapper._prepare_tokenizer_input(
             text,
             InputFormat.SINGLE_STRING,
         )
@@ -158,8 +156,7 @@ class TestTokenizerInputPreparation(unittest.TestCase):
     def test_prepare_batch_strings_input(self):
         """Test preparation of batch strings input."""
         texts = ["Hello", "World", "Test"]
-        result = TokenizerManager._prepare_tokenizer_input(
-            self.tokenizer_manager.raw_tokenizer_wrapper,
+        result = self.tokenizer_manager.raw_tokenizer_wrapper._prepare_tokenizer_input(
             texts,
             InputFormat.BATCH_STRINGS,
         )
@@ -168,8 +165,7 @@ class TestTokenizerInputPreparation(unittest.TestCase):
     def test_prepare_cross_encoder_pairs_input(self):
         """Test preparation of cross-encoder pairs input."""
         texts = [["query1", "doc1"], ["query2", "doc2"]]
-        result = TokenizerManager._prepare_tokenizer_input(
-            self.tokenizer_manager.raw_tokenizer_wrapper,
+        result = self.tokenizer_manager.raw_tokenizer_wrapper._prepare_tokenizer_input(
             texts,
             InputFormat.CROSS_ENCODER_PAIRS,
         )
@@ -178,8 +174,7 @@ class TestTokenizerInputPreparation(unittest.TestCase):
     def test_prepare_cross_encoder_single_pair_input(self):
         """Test preparation of single cross-encoder pair."""
         texts = [["query text", "document text"]]
-        result = TokenizerManager._prepare_tokenizer_input(
-            self.tokenizer_manager.raw_tokenizer_wrapper,
+        result = self.tokenizer_manager.raw_tokenizer_wrapper._prepare_tokenizer_input(
             texts,
             InputFormat.CROSS_ENCODER_PAIRS,
         )
@@ -188,8 +183,7 @@ class TestTokenizerInputPreparation(unittest.TestCase):
     def test_prepare_batch_strings_input_format_passthrough(self):
         """Batch strings should pass through unchanged."""
         texts = ["test"]
-        result = TokenizerManager._prepare_tokenizer_input(
-            self.tokenizer_manager.raw_tokenizer_wrapper,
+        result = self.tokenizer_manager.raw_tokenizer_wrapper._prepare_tokenizer_input(
             texts,
             InputFormat.BATCH_STRINGS,
         )
@@ -221,8 +215,7 @@ class TestTokenizerResultExtraction(unittest.TestCase):
         token_type_ids = [[0, 0, 0]]
 
         result_input_ids, result_token_type_ids = (
-            TokenizerManager._extract_tokenizer_results(
-                self.tokenizer_manager.raw_tokenizer_wrapper,
+            self.tokenizer_manager.raw_tokenizer_wrapper._extract_tokenizer_results(
                 input_ids,
                 token_type_ids,
                 InputFormat.SINGLE_STRING,
@@ -239,8 +232,7 @@ class TestTokenizerResultExtraction(unittest.TestCase):
         token_type_ids = [[0, 0, 0, 1, 1]]
 
         result_input_ids, result_token_type_ids = (
-            TokenizerManager._extract_tokenizer_results(
-                self.tokenizer_manager.raw_tokenizer_wrapper,
+            self.tokenizer_manager.raw_tokenizer_wrapper._extract_tokenizer_results(
                 input_ids,
                 token_type_ids,
                 InputFormat.CROSS_ENCODER_PAIRS,
@@ -257,8 +249,7 @@ class TestTokenizerResultExtraction(unittest.TestCase):
         token_type_ids = [[0, 0, 0], [0, 0, 0]]
 
         result_input_ids, result_token_type_ids = (
-            TokenizerManager._extract_tokenizer_results(
-                self.tokenizer_manager.raw_tokenizer_wrapper,
+            self.tokenizer_manager.raw_tokenizer_wrapper._extract_tokenizer_results(
                 input_ids,
                 token_type_ids,
                 InputFormat.BATCH_STRINGS,
@@ -275,8 +266,7 @@ class TestTokenizerResultExtraction(unittest.TestCase):
         token_type_ids = [[0, 0, 0, 1, 1], [0, 0, 0, 1, 1]]
 
         result_input_ids, result_token_type_ids = (
-            TokenizerManager._extract_tokenizer_results(
-                self.tokenizer_manager.raw_tokenizer_wrapper,
+            self.tokenizer_manager.raw_tokenizer_wrapper._extract_tokenizer_results(
                 input_ids,
                 token_type_ids,
                 InputFormat.CROSS_ENCODER_PAIRS,
@@ -295,8 +285,7 @@ class TestTokenizerResultExtraction(unittest.TestCase):
         token_type_ids = None
 
         result_input_ids, result_token_type_ids = (
-            TokenizerManager._extract_tokenizer_results(
-                self.tokenizer_manager.raw_tokenizer_wrapper,
+            self.tokenizer_manager.raw_tokenizer_wrapper._extract_tokenizer_results(
                 input_ids,
                 token_type_ids,
                 InputFormat.SINGLE_STRING,
@@ -313,8 +302,7 @@ class TestTokenizerResultExtraction(unittest.TestCase):
         token_type_ids = None
 
         result_input_ids, result_token_type_ids = (
-            TokenizerManager._extract_tokenizer_results(
-                self.tokenizer_manager.raw_tokenizer_wrapper,
+            self.tokenizer_manager.raw_tokenizer_wrapper._extract_tokenizer_results(
                 input_ids,
                 token_type_ids,
                 InputFormat.SINGLE_STRING,
@@ -350,14 +338,18 @@ class TestTokenizerManagerIntegration(unittest.TestCase):
         text = "Hello world"
 
         # Step 1: Detect format
-        input_format = TokenizerManager._detect_input_format(
-            self.tokenizer_manager.raw_tokenizer_wrapper, text, is_cross_encoder=False
+        input_format = (
+            self.tokenizer_manager.raw_tokenizer_wrapper._detect_input_format(
+                text, is_cross_encoder=False
+            )
         )
         self.assertEqual(input_format, InputFormat.SINGLE_STRING)
 
         # Step 2: Prepare input
-        tokenizer_input = TokenizerManager._prepare_tokenizer_input(
-            self.tokenizer_manager.raw_tokenizer_wrapper, text, input_format
+        tokenizer_input = (
+            self.tokenizer_manager.raw_tokenizer_wrapper._prepare_tokenizer_input(
+                text, input_format
+            )
         )
         self.assertEqual(tokenizer_input, ["Hello world"])
 
@@ -366,8 +358,7 @@ class TestTokenizerManagerIntegration(unittest.TestCase):
         mock_token_type_ids = None
 
         result_input_ids, result_token_type_ids = (
-            TokenizerManager._extract_tokenizer_results(
-                self.tokenizer_manager.raw_tokenizer_wrapper,
+            self.tokenizer_manager.raw_tokenizer_wrapper._extract_tokenizer_results(
                 mock_input_ids,
                 mock_token_type_ids,
                 input_format,
@@ -385,14 +376,18 @@ class TestTokenizerManagerIntegration(unittest.TestCase):
         ]
 
         # Step 1: Detect format
-        input_format = TokenizerManager._detect_input_format(
-            self.tokenizer_manager.raw_tokenizer_wrapper, texts, is_cross_encoder=True
+        input_format = (
+            self.tokenizer_manager.raw_tokenizer_wrapper._detect_input_format(
+                texts, is_cross_encoder=True
+            )
         )
         self.assertEqual(input_format, InputFormat.CROSS_ENCODER_PAIRS)
 
         # Step 2: Prepare input
-        tokenizer_input = TokenizerManager._prepare_tokenizer_input(
-            self.tokenizer_manager.raw_tokenizer_wrapper, texts, input_format
+        tokenizer_input = (
+            self.tokenizer_manager.raw_tokenizer_wrapper._prepare_tokenizer_input(
+                texts, input_format
+            )
         )
         self.assertEqual(tokenizer_input, texts)
 
@@ -401,8 +396,7 @@ class TestTokenizerManagerIntegration(unittest.TestCase):
         mock_token_type_ids = [[0, 0, 0, 0, 1, 1, 1]]
 
         result_input_ids, result_token_type_ids = (
-            TokenizerManager._extract_tokenizer_results(
-                self.tokenizer_manager.raw_tokenizer_wrapper,
+            self.tokenizer_manager.raw_tokenizer_wrapper._extract_tokenizer_results(
                 mock_input_ids,
                 mock_token_type_ids,
                 input_format,
@@ -418,14 +412,18 @@ class TestTokenizerManagerIntegration(unittest.TestCase):
         texts = ["Hello", "World", "Test"]
 
         # Step 1: Detect format
-        input_format = TokenizerManager._detect_input_format(
-            self.tokenizer_manager.raw_tokenizer_wrapper, texts, is_cross_encoder=False
+        input_format = (
+            self.tokenizer_manager.raw_tokenizer_wrapper._detect_input_format(
+                texts, is_cross_encoder=False
+            )
         )
         self.assertEqual(input_format, InputFormat.BATCH_STRINGS)
 
         # Step 2: Prepare input
-        tokenizer_input = TokenizerManager._prepare_tokenizer_input(
-            self.tokenizer_manager.raw_tokenizer_wrapper, texts, input_format
+        tokenizer_input = (
+            self.tokenizer_manager.raw_tokenizer_wrapper._prepare_tokenizer_input(
+                texts, input_format
+            )
         )
         self.assertEqual(tokenizer_input, ["Hello", "World", "Test"])
 
@@ -434,8 +432,7 @@ class TestTokenizerManagerIntegration(unittest.TestCase):
         mock_token_type_ids = None
 
         result_input_ids, result_token_type_ids = (
-            TokenizerManager._extract_tokenizer_results(
-                self.tokenizer_manager.raw_tokenizer_wrapper,
+            self.tokenizer_manager.raw_tokenizer_wrapper._extract_tokenizer_results(
                 mock_input_ids,
                 mock_token_type_ids,
                 input_format,
